requirements.txt
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.0.3
google-auth==2.37.0
idna==3.10
iniconfig==2.0.0
kubernetes==18.20.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytest==7.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tomli==2.2.1
urllib3==2.3.0
websocket-client==1.8.0
